Stomach cancer: If you also have these symptoms then be alert
Cancer does not always give symptoms. Many times, cancer goes undiagnosed, that is, it shows little or sometimes no signs or symptoms, making it difficult to detect in the early stages.

There are other cases in which the symptoms produced by cancer are similar to those of many routine infections or conditions, often resulting in misdiagnosis and late detection. One such cancer is stomach or gastric cancer, which remains undiagnosed in its initial stages and later reaches the advanced stage.

The stomach is in the middle of our intestines and is anatomically related to other nearby structures and organs such as the liver, pancreas, esophagus, and colon. For this reason, these structures share common risk factors for the development of cancer. Most stomach cancers are adenocarcinomas that form in the glandular cells of the innermost lining of the stomach. Due to the structural and functional condition of the stomach, symptoms resulting from stomach cancer are often overlooked. If you are looking <a href="https://ujalacygnus.com/haldwani/">best hospital in haldwani</a> for cancer treatment, Ujala Cygnus Central hospital is the best oncology hospital in kumaon region.

Based on its appearance, the symptoms of colon cancer can be classified into two broad types – early stage and advanced stage symptoms.

early stage symptoms
People do not see any symptoms in stomach cancer in the initial stage. But, as the cancer starts to grow and spread, its symptoms can be felt like: Abdominal pain above the navel
heartburn
feeling full after eating only a small amount of food
frequent bouts of indigestion
nausea and vomiting
loss of appetite
rapid weight loss

When you read about the symptoms, you come to know that many of these symptoms are also seen in other conditions like bowel syndrome, GERD, food allergies, stomach ulcers, viral or bacterial infections etc. Many of these are treated by people with home remedies, very few people go to the doctor when such symptoms arise, which is not correct.

advanced stage symptoms
Gastric cancer is the fourth most common cancer in India as compared to western countries and it is the third most common cancer among men.
If it is not detected in the early stages, then it reaches the advanced stage and can also spread to other organs like liver, lungs and perineum. The symptoms experienced by patients in the advanced stage are far more severe than those in the early stages.

These symptoms include:

body weakness and exhaustion
blood in stool
uncontrolled vomiting
jaundice when cancer affects the liver

Swelling or fluid formation in the bowel Colon cancer usually reaches the advanced stage when it is not regularly screened. However, in countries such as Japan, where colon cancer is common, large-scale screening is done regularly, resulting in early detection of the condition. However, routine screening has not yet been proven to be beneficial.